pyproject.toml
[project]
name = "portfolio-optimize-assistant"
version = "0.1.0"
description = "Agentic portfolio optimization engine for structured notes"
requires-python = ">=3.11"

dependencies = [
    # Agent Orchestration
    "langgraph>=1.0.0,<2.0",
    "langchain-core>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-community>=0.3.0",

    # LLM
    "anthropic>=0.40.0",

    # PDF Parsing
    "pdfplumber>=0.11.0",
    "PyMuPDF>=1.24.0",

    # Vector Store + Embeddings
    "faiss-cpu>=1.8.0",
    "sentence-transformers>=3.0.0",

    # Optimization
    "cvxpy>=1.5.0",
    "osqp>=0.6.0",

    # Market Data
    "yfinance>=0.2.40",

    # Data
    "numpy>=1.26.0,<2.1",
    "pandas>=2.1.0",
    "pydantic>=2.5.0",

    # UI
    "streamlit>=1.38.0",

    # Utilities
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.5.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"
